Remove onCancel from useAccountInfo hook
